Extract `DetermineMinIndentation` and `IsSpace` to the helpers package
