+++ format/heredoc.go
@@ -26,35 +26,3 @@ func removeIndentation(lines []string, n int) []string {
 	}
 	return lines
 }
-
-// Checks if the given rune is a space character.
-func isSpace(r rune) bool {
-	return r == ' ' || r == '\t'
-}
-
-// determineMinIndentation determines the minimum indentation level of the given lines.
-func determineMinIndentation(lines []string) int {
-	// Initialize minimum indentation level
-	min := 0
-
-	// Iterate over lines and determine indentation level
-	for _, line := range lines {
-		indentationLevel := 0 // number of spaces at the beginning of the line
-		// Determine indentation level
-		for _, r := range line {
-			if isSpace(r) {
-				indentationLevel++
-			} else {
-				break
-			}
-		}
-
-		// Update minimum indentation level
-		if indentationLevel < min || min == 0 {
-			min = indentationLevel
-		}
-	}
-
-	// Return minimum indentation level
-	return min
-}
